betty format

+++ singly_linked_lists/0-print_list.c
@@ -9,19 +9,18 @@
  */
 size_t print_list(const list_t *h)
 {
-	size_t count = 0;
+	unsigned int count = 0;
 
 	while (h != NULL)
 	{
 		if (h->str == NULL)
-		printf("[0] (nil)\n");
-
+			printf("[0] (nil)\n");
 		else
-		printf("[%u] %s\n", h->len, h->str);
+			printf("[%u] %s\n", h->len, h->str);
 
 		h = h->next;
 		count++;
 	}
 
-return (count);
+	return (count);
 }
